Remove page-size input
